More clippy fixes.
